requirements.txt
adbutils==1.0.8
airtest==1.2.7
allure-pytest==2.11.0
allure-python-commons==2.11.0
apkutils2==1.0.0
Appium-Python-Client==2.7.0
async-generator==1.10
attrs==22.1.0
cached-property==1.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cigam==0.0.3
click==8.1.3
colored==1.4.4
coloredlogs==15.0.1
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
exceptiongroup==1.0.0rc9
facebook-wda==1.4.6
filelock==3.8.0
fire==0.4.0
Flask==2.1.3
Flask-SocketIO==4.3.1
flatbuffers==22.10.26
gevent==22.10.1
greenlet==1.1.3.post0
h11==0.14.0
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.0.1
loguru==0.6.0
logzero==1.7.0
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
mpmath==1.2.1
mss==6.1.0
numpy==1.23.5
onnxruntime==1.13.1
opencv-contrib-python==4.6.0.66
outcome==1.2.0
packaging==20.9
Pillow==9.3.0
pluggy==1.0.0
progress==1.6
protobuf==4.21.9
py==1.11.0
pyelftools==0.29
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-rerunfailures==10.2
python-engineio==3.13.2
python-poco==0.1.0
python-socketio==4.6.0
python-xlib==0.32
pywinauto==0.6.3
requests==2.28.1
retry==0.9.2
selenium==4.5.0
simple-tornado==0.2.2
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
solox==2.3.5
sortedcontainers==2.4.0
sympy==1.11.1
tabulate==0.9.0
termcolor==2.1.1
tidevice==0.9.11
tomli==2.0.1
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
uiautomator2==2.16.19
urllib3==1.26.12
websocket==0.2.1
websocket-client==1.4.1
weditor==0.6.7
Werkzeug==2.0.3
whichcraft==0.6.1
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.0
